requirements.txt
fastapi 
uvicorn
sqlalchemy
psycopg2
posslib
pyjwt
email_validator
bcrypt
jinja2
python-multipart
httpx